[react-addons-test-utils] Update definitions for 15.6
