+++ types/react-addons-test-utils/index.d.ts
@@ -1,4 +1,4 @@
-// Type definitions for React (react-addons-test-utils) 0.14
+// Type definitions for React (react-addons-test-utils) 15.6
 // Project: http://facebook.github.io/react/
 // Definitions by: Asana <https://asana.com>, AssureSign <http://www.assuresign.com>, Microsoft <https://microsoft.com>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
