+++ types/react-addons-test-utils/tsconfig.json
@@ -1,6 +1,7 @@
 {
     "files": [
-        "index.d.ts"
+        "index.d.ts",
+        "react-addons-test-utils-tests.ts"
     ],
     "compilerOptions": {
         "module": "commonjs",
@@ -17,6 +18,11 @@
             "../"
         ],
         "types": [],
+        "paths": {
+            "react": [
+                "react/v15"
+            ]
+        },
         "noEmit": true,
         "forceConsistentCasingInFileNames": true
     }
